Using <JSP:PARAM> Tag
In this project, we'll be working with the <JSP:PARAM> tag. Think of it as the "include" function's more versatile cousin. Not only can it include content, but it also allows you to override data by adding new values. It's like having a superpower that lets you tweak and customize your data on the fly..!