Form and MySQL Database Integration
This project is quite the adventure! For the first time, we'll connect our form to a MySQL database. Not only will we store our data, but we can also edit and delete it. Think of it as having your very own data playground where you can add, tweak, and clean up your information as you see fit!